not pass usertests

+++ user/user.h
@@ -23,7 +23,7 @@ int getpid(void);
 char* sbrk(int);
 int sleep(int);
 int uptime(void);
-void *mmap(void *,size_t ,int,int,int,off_t);
+void *mmap(void *,size_t ,int,int,int,uint);
 int munmap(void *,size_t);
